% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{clst}
\alias{clst}
\title{Exemplary hsa_quant class object}
\format{
a hsa_quant class object
}
\usage{
clst
}
\description{
clst <- hsa_quant(hsa_rez)
usethis::use_data(clst)
}
\keyword{datasets}
