% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{density_hespdiv}
\alias{density_hespdiv}
\title{hespdiv results of density based subdivision}
\format{
a hespdiv object
}
\usage{
density_hespdiv
}
\description{
derivation is in demo.R
}
\keyword{datasets}
