% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{group_contributions}
\alias{group_contributions}
\title{family contributions to splits in hd_rcase}
\format{
a group_effect object
}
\usage{
group_contributions
}
\description{
derivation is in demo.R
}
\keyword{datasets}
