% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{h_rez}
\alias{h_rez}
\title{hespdiv object for workshop (simulated two bioregions)}
\format{
a hespdiv class object
}
\usage{
h_rez
}
\description{
derivation is in demo.R
}
\keyword{datasets}
