% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{hd}
\alias{hd}
\title{a hespdiv class object obtained with the default hespdiv arguments and package datasets}
\format{
a hespdiv class object
}
\usage{
hd
}
\description{
species <- mio_mams$accepted_name # taxa names
sp_coords <- data.frame(x = mio_mams$lng, y = mio_mams$lat)
hd <- hespdiv(data = species, xy.dat = sp_coords, study.pol = us)
usethis::use_data(hd)
}
\keyword{datasets}
