% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{hd_rcase}
\alias{hd_rcase}
\title{hespdiv object for workshop (like in paper)}
\format{
a hespdiv class object
}
\usage{
hd_rcase
}
\description{
derivation is in demo.R
}
\keyword{datasets}
