% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDData-package.R
\docType{package}
\name{HDData-package}
\alias{HDData}
\alias{HDData-package}
\title{HDData: Data Repository for hespdiv Package}
\description{
Provides exemplary datasets for hespdiv package.
}
\author{
\strong{Maintainer}: Liudas Daumantas \email{liudas.daumantas@chgf.vu.lt}

}
\keyword{internal}
