% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{hsa_rez}
\alias{hsa_rez}
\title{Examplary HespDiv sensitivity analysis object obtained by trying alternative
arguments for hd run.}
\format{
a hsa class object with 100 alternative hespdiv runs
}
\usage{
hsa_rez
}
\description{
set.seed(2) # seed is used to obtained the same result of an experiment with random properties.
hsa_rez <- hsa(obj = hd,
n.runs = 100, # 100 alternative hespdiv re-reruns
n.split.pts = 8:30, # number of split-points determines fit to data of straight split-lines
same.n.split = FALSE, # split-point placement regularity determines whether scale of analysis changes depending on order of subdivision
c.splits = FALSE,  # Argument controls whether curves are generated in attempt to increase the performance of the best linear split-lines
c.X.knots = 3:8, # Controls the number of wiggles in generated curves, determining their fit to data
c.Y.knots = 5:15, # Controls the number of different shapes each curve wiggle can achieve, also determining their fit to data
c.fast.optim = TRUE, # Determines the optimization algorithm of non-linear split-lines
use.chull = FALSE) # Determines whether the convex polygon of occurences is used as a study area polygon. If not, study area polygon becomes the provided US polygon.
}
\keyword{datasets}
