% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{mio_mams}
\alias{mio_mams}
\title{Miocene mammal occurrences in US, Palaeobiology database data}
\format{
a data frame of size 5244x132
}
\usage{
mio_mams
}
\description{
This data set was obtained using data_prep.R script. Its a sixth object
of us_l (\code{us_l[[6]]})
It is a processed data set of Miocene mammal occurrences in US downloaded from
the Palaeobiology database.
Observations with poor location precision were removed, along with marine and flying mammals.
Also, 8 observations were removed because they lie outside US polygon.
Raw data was downloaded 2023 April 12.
}
\keyword{datasets}
