% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{nl}
\alias{nl}
\title{A nullhespdiv class object obtain with the default arguments and package data}
\format{
a nullhespdiv class object
}
\usage{
nl
}
\description{
set.seed(1)
nl <- nulltest(hd, n = 1000)
usethis::use_data(nl)
}
\keyword{datasets}
