% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{simpson_hespdiv}
\alias{simpson_hespdiv}
\title{hespdiv results of subdivision based on linear relationship differences}
\format{
a hespdiv object
}
\usage{
simpson_hespdiv
}
\description{
derivation is in demo.R
}
\keyword{datasets}
