% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{species_contributions}
\alias{species_contributions}
\title{species contributions to splits in hd_rcase}
\format{
a taxon_effect_result object
}
\usage{
species_contributions
}
\description{
derivation is in demo.R
}
\keyword{datasets}
