% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa_rez.R
\docType{data}
\name{us}
\alias{us}
\title{Polygon of the Contiguous United States of America.}
\format{
a data frame of size 233x2
}
\usage{
us
}
\description{
Was downloaded 2023 April 12 from spData package, data(world)
}
\keyword{datasets}
